Add stop timer functionality for discordWrapper
The Rich Presence activity does not get changed until another Toggl timer
replaces the previous one. If timers are stopped and not replaced, the old
Rich Presence activity continues to go on even if the original Toggl timer
has been stopped.

Replace previous discordWrapper refresh function to restart the Rich Presence
client if the Toggl timer is stopped and not replaced. Fix up original
CurrentTimer function to also pass errors in case of empty timer.

Do note that "rich-go", the library used for the Discord RP implementation, does
not have an option to output an empty activity (or nil of some kind), so
restarting is necessary.