Add "refresh interval" configuration to CLI
By default, TogglRichPresence refreshes the Rich Presence activity every 60
seconds. Add environment variable support to change this configuration. Default
new refresh interval is 10 seconds, as this is more sensible to usual changes in
timer entries, and it is not too frequent to rate-limit the Toggl API.

Format and lint previously unformatted files.